Create Custom Exceptions

+++ CongestionTaxCalculator.Application/Features/Rules/Handlers/Commands/DeleteRuleCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using CongestionTaxCalculator.Application.Exceptions;
 using CongestionTaxCalculator.Application.Features.Rules.Requests.Commands;
 using CongestionTaxCalculator.Application.Persistence.Contracts;
 using MediatR;
@@ -17,6 +18,10 @@ public class DeleteRuleCommandHandler : IRequestHandler<DeleteRuleCommand,Unit>
     public async Task<Unit> Handle(DeleteRuleCommand request, CancellationToken cancellationToken)
     {
         var rule = await _ruleRepository.GetById(request.Id);
+        if(rule == null)
+        {
+            throw new NotFoundException(nameof(rule), request.Id);
+        }
         await _ruleRepository.Delete(rule);
         return Unit.Value;
     }
